% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genotype_sp_extraction.R
\name{Genotype_sp_extraction}
\alias{Genotype_sp_extraction}
\title{Genotype extraction and filtering for association analysis}
\usage{
Genotype_sp_extraction(
  genofile,
  variant.id,
  sample.id,
  REF_AF,
  rare_maf_cutoff = 0.01,
  variant_maf_cutoff_filter = 1,
  Missing_rate,
  Missing_cutoff = 0.01,
  subset_variants_num = 2000,
  rv_num_cutoff_max_prefilter = 1e+09,
  annotation_phred = NULL
)
}
\arguments{
\item{genofile}{an object of opened annotated GDS (aGDS) file.}

\item{variant.id}{ID of selected variants.}

\item{sample.id}{ID of selected samples.}

\item{REF_AF}{a numeric vector of reference allele frequencies for each variant.}

\item{rare_maf_cutoff}{the cutoff of maximum minor allele frequency in defining rare variants (default = 0.01).}

\item{variant_maf_cutoff_filter}{a numeric value specifying the MAF threshold for excluding common variants in variant-set analysis. Default is 1 for individual-variant analysis.}

\item{Missing_rate}{a numeric vector of missing rates for each variant.}

\item{Missing_cutoff}{a numeric value specifying the maximum missing rate threshold for defining high-missing variants (default = 0.01).}

\item{subset_variants_num}{the number of variants to extract per subset for each time in Case 1 and Case 2 (default = 2e3).}

\item{rv_num_cutoff_max_prefilter}{the cutoff of maximum number of variants before extracting the genotype matrix (default = 1e+09).}

\item{annotation_phred}{a optional data frame or matrix of functional annotation data of dimension p*q (or a vector of a single annotation score with length p). See \code{STAAR::STAAR} for more details.}
}
\value{
A list containing:
\item{Geno}{A sparse matrix of genotypes.}
\item{results_information}{A data frame with variant-level details, including chromosome, position, reference and alternative alleles, MAF, ALT_AF, missing rate, and variant ID.}
\item{annotation_phred}{A data frame containing filtered functional annotation data if provided; otherwise, NULL.}
}
\description{
The \code{Genotype_sp_extraction} function extracts genotype data for a given set of variants and samples from a GDS file.
It applies multiple filters based on allele frequency, missingness rate, and user-defined thresholds to categorize variants into different dosage groups.
The function returns a sparse genotype matrix (\code{"dgCMatrix"}), variant-level information, and optional annotation data.
}
\details{
The extracted variants are processed based on the following three cases:
(1) Case 1: \code{ALT_AF > 0.5}
    - Use \code{"$dosage"} to extract Geno (dosages of the reference allele).
    - Convert the extracted data into the \code{"dgCMatrix"} format (sparse matrix).

(2) Case 2: \code{ALT_AF ≤ 0.5} and (\code{MAF ≥ rare_maf_cutoff} or \code{Missing_rate ≥ Missing_cutoff})
    - Use \code{"$dosage_alt"} to extract Geno (dosages of the alternative allele).
    - Convert the extracted data into the \code{"dgCMatrix"} format (sparse matrix).

(3) Case 3: \code{ALT_AF ≤ 0.5} and (\code{MAF < rare_maf_cutoff} and \code{Missing_rate < Missing_cutoff})
    - Use \code{"$dosage_ap"} to directly extract Geno in the \code{"dgCMatrix"} format.

Note:
- \code{REF_AF} and \code{Missing_rate} can be efficiently computed using
  \code{SeqArray::seqGetAF_AC_Missing(genofile, minor=FALSE)}.
}
